Fix build by moving mock-data and styles to assets
